+++ src/pages/signup/ui/signup.tsx
@@ -1,32 +1,21 @@
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useRef, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useNavigate } from "react-router";
 import { ROUTE_PATH } from "@/app/routes/paths";
-import { useSignUp } from "@/pages/signup/apis/mutations/use-signup";
 import {
 	type SignupFormData,
 	type SignupFormInput,
 	signupSchema,
 } from "@/pages/signup/model/signup-schema";
-import { requestKakaoAuthorize } from "@/shared/libs/request-kakao-authorize";
 import { Button } from "@/shared/ui/buttons/button";
-import { CheckBox } from "@/shared/ui/check-box/check-box";
 import { DateInput } from "@/shared/ui/inputs/date-input";
 import { InputMedium } from "@/shared/ui/inputs/input-medium";
 import { CategoryLabel } from "@/shared/ui/labels/category-label";
 import { Header } from "@/shared/ui/navigations/header";
-import { openModal } from "@/shared/ui/overlays/modal/open-modal";
-import { notifyError } from "@/shared/ui/overlays/toast/toast";
 import { RadioButton } from "@/shared/ui/radio/radio";
 
 export const Signup = () => {
 	const navigate = useNavigate();
-	const [isAgreed, setIsAgreed] = useState(false);
-	const [isCheckboxEnabled, setIsCheckboxEnabled] = useState(false);
-	const { mutate } = useSignUp();
-
-	const agreementSectionRef = useRef<HTMLElement>(null);
 
 	const {
 		register,
@@ -58,83 +47,12 @@ export const Signup = () => {
 		birthdate.day !== "" &&
 		gender !== undefined;
 
-	// 회원가입 완료 모달 열기
-	const openSignupCompleteModal = () => {
-		openModal({
-			size: "sm",
-			title: "회원가입이 완료되었어요",
-			description:
-				"이미 건강 검진을 받아보셨다면,\n결과를 계속해서 입력할까요?",
-			secondaryAction: {
-				label: "메인으로 가기",
-				onClick: () => {
-					localStorage.setItem("signupRedirect", "home");
-					requestKakaoAuthorize();
-				},
-			},
-			primaryAction: {
-				label: "이어서 입력하기",
-				onClick: () => {
-					localStorage.setItem("signupRedirect", "checkup-result");
-					requestKakaoAuthorize();
-				},
-			},
-		});
-	};
-
 	const onSubmit = (data: SignupFormData) => {
-		if (!isAgreed) {
-			agreementSectionRef.current?.scrollIntoView({
-				behavior: "smooth",
-				block: "center",
-			});
-			notifyError("개인정보 수집·이용에 동의해주세요");
-			return;
-		}
-
-		mutate(
-			{
+		void navigate(ROUTE_PATH.SIGNUP_TOS, {
+			state: {
 				name: data.name,
 				birthdate: `${data.birthdate.year}-${data.birthdate.month.padStart(2, "0")}-${data.birthdate.day.padStart(2, "0")}`,
 				gender: data.gender,
-			},
-			{
-				onSuccess: () => {
-					openSignupCompleteModal();
-				},
-				onError: () => {
-					void navigate(ROUTE_PATH.LOGIN, { replace: true });
-					notifyError("회원가입에 실패했습니다");
-				},
-			},
-		);
-	};
-
-	// 개인정보 수집·이용 모달 열기
-	const handleOpenPrivacyModal = () => {
-		openModal({
-			title: "개인정보 수집·이용 동의",
-			description: `케어나(이하 '서비스')는 이용자의 건강 정보를 안전하게 보호하기 위해 아래와 같이 개인정보를 수집·이용합니다.
-
-1. 수집하는 개인정보 항목
-- 필수 항목: 이름, 생년월일, 성별, 건강검진 결과(검진일자, 검진기관, 검사 수치 등)
-
-2. 개인정보의 수집·이용 목적
-- 건강검진 결과 분석 및 맞춤형 건강 정보 제공
-- 서비스 이용에 따른 본인 확인
-- 서비스 개선 및 신규 서비스 개발
-
-3. 개인정보의 보유 및 이용 기간
-- 회원 탈퇴 시까지 또는 동의 철회 시까지
-- 단, 관계 법령에 따라 보존이 필요한 경우 해당 기간 동안 보관
-
-4. 동의 거부권 및 불이익
-- 이용자는 개인정보 수집·이용에 대한 동의를 거부할 권리가 있습니다.
-- 다만, 필수 항목에 대한 동의를 거부하실 경우 서비스 이용이 제한됩니다.`,
-			onScrollEnd: () => setIsCheckboxEnabled(true),
-			primaryAction: {
-				label: "확인",
-				onClick: () => {},
 			},
 		});
 	};
@@ -239,50 +157,14 @@ export const Signup = () => {
 				</section>
 			</form>
 
-			{/* TODO: 공컴 추출 */}
-			{/* 하단 영역 (개인정보 동의 + 버튼) */}
-			<div className="flex flex-col">
-				{/* 개인정보 동의 - 기본정보 입력 완료 시 표시 */}
-				{isRequiredFilled && isValid && (
-					<section
-						ref={agreementSectionRef}
-						className="fade-in-animation mb-[3.6rem] flex flex-col gap-[2rem]"
-					>
-						<h3 className="head02-b-16 text-gray-600">
-							개인정보 수집·이용 동의
-						</h3>
-						<div className="flex items-start gap-[0.8rem]">
-							<CheckBox
-								checked={isAgreed}
-								onChange={setIsAgreed}
-								disabled={!isCheckboxEnabled}
-							/>
-							<p className="whitespace-nowrap pt-[0.2rem] text-black">
-								<span className="body01-sb-12">[필수] </span>
-								<button
-									type="button"
-									className="body01-sb-12 underline underline-offset-2"
-									onClick={handleOpenPrivacyModal}
-								>
-									개인정보 수집·이용
-								</button>{" "}
-								<span className="body05-r-12">
-									내용을 확인하였으며 이에 동의합니다.
-								</span>
-							</p>
-						</div>
-					</section>
-				)}
-
-				<Button
-					type="submit"
-					form="signup-form"
-					size="lg"
-					disabled={!isRequiredFilled || !isValid}
-				>
-					회원가입
-				</Button>
-			</div>
+			<Button
+				type="submit"
+				form="signup-form"
+				size="lg"
+				disabled={!isRequiredFilled || !isValid}
+			>
+				확인
+			</Button>
 		</div>
 	);
 };
